+++ CHANGES.rst
@@ -13,6 +13,13 @@ Improvements
 - Hide category field in event creation dialog if there are no
   subcategories (:issue:`3112`)
 
+Bugfixes
+^^^^^^^^
+
+- Do not intercept HTTP exceptions containing a custom response.
+  When raising such exceptions we do not want the default handling
+  but rather send the custom response to the client.
+
 
 Version 2.0a1
 -------------